factored __assemble_ bin scripts into bin/shared/packageAssembly_utils.node.js; including a few asset directories to filter from browser build; filtering Lite files from cordova and electron packages
